+++ provider/pro/testdata/pro_database_auto_minor_version_upgrade.tf
@@ -1,7 +1,7 @@
 locals {
   rediscloud_cloud_account     = "%s"
   rediscloud_subscription_name = "%s"
-  rediscloud_password          = "%s"
+  auto_minor_version_upgrade = %s
 }
 
 data "rediscloud_payment_method" "card" {
@@ -14,6 +14,7 @@ data "rediscloud_cloud_account" "account" {
   provider_type            = "AWS"
   name                     = local.rediscloud_cloud_account
 }
+
 resource "rediscloud_subscription" "example" {
   name              = local.rediscloud_subscription_name
   payment_method_id = data.rediscloud_payment_method.card.id
@@ -27,43 +28,23 @@ resource "rediscloud_subscription" "example" {
       preferred_availability_zones = ["eu-west-1a"]
     }
   }
+
   creation_plan {
     memory_limit_in_gb           = 1
     quantity                     = 1
     replication                  = false
-    support_oss_cluster_api      = true
     throughput_measurement_by    = "operations-per-second"
     throughput_measurement_value = 1000
-    modules = ["RediSearch"]
   }
 }
+
 resource "rediscloud_subscription_database" "example" {
   subscription_id              = rediscloud_subscription.example.id
-  name                         = "tf-database"
+  name                         = "auto-minor-version-upgrade-test"
   protocol                     = "redis"
   memory_limit_in_gb           = 1
   data_persistence             = "none"
   throughput_measurement_by    = "operations-per-second"
   throughput_measurement_value = 1000
-  password                     = local.rediscloud_password
-  support_oss_cluster_api      = true
-  replication                  = false
-  enable_default_user          = true
-  query_performance_factor     = "2x"
-  redis_version                = "7.4"
-  modules = [
-    {
-      name : "RediSearch"
-    }
-  ]
-}
-
-data "rediscloud_database" "example-by-id" {
-  subscription_id = rediscloud_subscription.example.id
-  db_id           = rediscloud_subscription_database.example.db_id
-}
-
-data "rediscloud_database" "example-by-name" {
-  subscription_id = rediscloud_subscription.example.id
-  name            = rediscloud_subscription_database.example.name
+  auto_minor_version_upgrade   = local.auto_minor_version_upgrade
 }